requirements.txt
pandas
django
numpy
djangorestframework
pysqlite3
sqlalchemy
UPath
requests
UPath
pymysql
pyarrow
